OAK-3158 AE when specifiying 2G cache for FileStore
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1702240 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/blob/datastore/DataStoreBlobStore.java
@@ -109,7 +109,7 @@ public class DataStoreBlobStore implements DataStore, SharedDataStore, BlobStore
 
         this.cache = CacheLIRS.<String, byte[]>newBuilder()
                 .module("DataStoreBlobStore")
-                .maximumWeight(cacheSizeInMB * FileUtils.ONE_MB)
+                .maximumWeight((long) cacheSizeInMB * FileUtils.ONE_MB)
                 .weigher(new Weigher<String, byte[]>() {
                     @Override
                     public int weigh(@Nonnull String key, @Nonnull byte[] value) {
